fix(Space): 垂直间距补偿值计算问题修复
