+++ components/space/const.ts
@@ -1,0 +1,5 @@
+import getPrefixCls from '../_util/getPrefixCls';
+
+export const COMPONENT_NAME = 'FSpace';
+
+export const prefixCls = getPrefixCls('space');
