+++ components/space/index.ts
@@ -5,8 +5,8 @@ import type { SFCWithInstall } from '../_util/interface';
 
 type SpaceType = SFCWithInstall<typeof Space>;
 
-export { spaceProps } from './space';
-export type { SpaceProps } from './space';
+export { spaceProps } from './props';
+export type { SpaceProps } from './props';
 export const FSpace = withInstall<SpaceType>(Space as SpaceType);
 
 export default FSpace;
